fix bug

+++ 100-shell_sort.c
@@ -10,7 +10,7 @@ void shell_sort(int *array, size_t size)
 {
 	int temp;
 	size_t i, j;
-	int interval = 1;
+	size_t interval = 1;
 
 	if (array == NULL || size < 2)
 		return;
